Display Table Enhancements
Add max_rows and include_index arguments to the table display function. The max_rows argument should impose a limit on the number of rows that can be displayed by the function to prevent accidental large table display issues. The include_index argument should allow the user to display the index column in the output display if they desire.