Step Stopping
Enhance step stopping to default to a non-silent mode that shows the
step was stopped in the display. Explicitly setting the stop will
avoid this display notification. Allow for setting a custom message in
the stop display as well.

Refactor templating and web assets to adjust to recent changes in SCSS
compilation and minification introduced in gulp plugins.