PyTest Support
This PR adds support for the `pytest` framework when running step tests with a new `@CauldronTest` decorator that wraps around a `pytest` function to handle the setup and teardown for each test. The same functionality available in the `StepTestCase` class for the `unittest` framework has been ported to the `@CauldronTest` decorator. A `cauldron.steptest.support` module has been added to hold shared functionality between the two test classes.