Step Testing Enhancements
This PR improves step testing by allowing the `run_step` method on the
StepTest class-based and function test objects to accept partial step
names. This allows a bit more flexibility in defining the step being
run and reduces the refactoring when steps are renamed. For example,
it is now possible for `run_step('foo')` to work in a project with
a step named `S05-foo.py`. The prefix and suffix can be ignored
when specifying the step. Although partial matching can lead to issues
if names are not distinct within a project.